Move fill-color field outside and refactor
