+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/fill-color-by-value-view.js
@@ -1,20 +1,25 @@
 var _ = require('underscore');
 var CoreView = require('backbone/core-view');
-var FillDialogModel = require('builder/components/form-components/editors/fill/fill-dialog-model');
-var FillDialogView = require('builder/components/form-components/editors/fill/fill-dialog');
 
-var InputCollection = require('builder/components/form-components/editors/fill/fill-color/fill-color-input-collection');
-var InputColorSolidView = require('builder/components/form-components/editors/fill/fill-color/inputs/input-color-solid');
-var InputImageView = require('builder/components/form-components/editors/fill/fill-color/inputs/input-image');
+var InputCollection = require('builder/components/form-components/editors/fill-color/fill-color-input-collection');
+var InputColorByValueView = require('builder/components/form-components/editors/fill-color/inputs/input-color-by-value');
 
 var FillTemplate = require('builder/components/form-components/editors/fill/fill-template.tpl');
+var FillDialogModel = require('builder/components/form-components/editors/fill/fill-dialog-model');
+var FillDialogView = require('builder/components/form-components/editors/fill/fill-dialog');
 var PopupManager = require('builder/components/popup-manager');
 
 var checkAndBuildOpts = require('builder/helpers/required-opts');
 
+var QUANTIFICATION_REF = {
+  'Jenks': 'jenks',
+  'Equal Interval': 'equal',
+  'Heads/Tails': 'headtails',
+  'Quantile': 'quantiles'
+};
+
 var INPUT_TYPE_MAP = {
-  image: InputImageView,
-  color: InputColorSolidView
+  color: InputColorByValueView
 };
 
 var REQUIRED_OPTS = [
@@ -25,20 +30,12 @@ var REQUIRED_OPTS = [
   'editorAttrs',
   'modals',
   'dialogMode',
-  'colorAttributes'
+  'colorAttributes',
+  'popupConfig'
 ];
 
 module.exports = CoreView.extend({
   className: 'Form-InputFill CDB-OptionInput CDB-Text js-input',
-
-  events: {
-    focus: function () {
-      this.trigger('focus', this);
-    },
-    blur: function () {
-      this.trigger('blur', this);
-    }
-  },
 
   initialize: function (options) {
     checkAndBuildOpts(options, REQUIRED_OPTS, this);
@@ -64,65 +61,11 @@ module.exports = CoreView.extend({
     this._initFillDialog();
     this._initInputFields();
 
-    this._popupManager = new PopupManager(this.cid, this.$el, this._fillDialogView.$el);
-  },
-
-  _initInputFields: function () {
-    var colorAttributes = this._colorAttributes;
-
-    this._inputCollection = new InputCollection();
-    this._inputCollection.add(_.extend({ type: 'image' }, colorAttributes));
-    this._inputCollection.add(_.extend({ type: 'color' }, colorAttributes));
-
-    this._inputCollection.each(function (inputModel) {
-      var type = inputModel.get('type');
-
-      var InputTypeView = INPUT_TYPE_MAP[type];
-
-      if (!InputTypeView) {
-        throw new Error(type + ' is not a valid type of constructor');
-      }
-
-      var inputTypeView = new InputTypeView({
-        model: inputModel,
-        columns: this.options.columns,
-        query: this.options.query,
-        configModel: this.options.configModel,
-        userModel: this.options.userModel,
-        modals: this.options.modals,
-        editorAttrs: this.options.editorAttrs ? this.options.editorAttrs[type] : {},
-        disabled: this.options.editorAttrs && this.options.editorAttrs.disabled
-      });
-
-      inputTypeView.bind('click', this._onInputClick, this);
-      this.$('.js-content').append(inputTypeView.render().$el);
-    }, this);
-
-    this._inputCollection.bind('onInputChanged', this._onInputChanged, this);
-  },
-
-  _onInputClick: function (inputModel) {
-    if (inputModel.get('selected')) {
-      this.removeDialog();
-      return;
-    }
-
-    inputModel.set('selected', true);
-    this._fillDialogView.model.set('createContentView', inputModel.get('createContentView'));
-    this._fillDialogView.render();
-    this._fillDialogView.show();
-
-    this._popupManager.append(this.dialogMode);
-    this._popupManager.track();
-  },
-
-  _onInputChanged: function (model) {
-    this._adjustImageSize(model);
-    this.trigger('onInputChanged', this);
-  },
-
-  _adjustImageSize: function (model) {
-    // TODO
+    this._popupManager = new PopupManager(
+      this._popupConfig.cid,
+      this._popupConfig.$el,
+      this._fillDialogView.$el
+    );
   },
 
   _initFillDialog: function () {
@@ -138,20 +81,67 @@ module.exports = CoreView.extend({
     });
   },
 
+  _initInputFields: function () {
+    var colorAttributes = this._colorAttributes;
+
+    this._inputCollection = new InputCollection();
+    this._inputCollection.add(_.extend({ type: 'color' }, colorAttributes));
+
+    this._inputCollection.each(function (inputModel) {
+      var type = inputModel.get('type');
+
+      var InputTypeView = INPUT_TYPE_MAP[type];
+
+      if (!InputTypeView) {
+        throw new Error(type + ' is not a valid type of constructor');
+      }
+
+      if (inputModel.get('quantification') && QUANTIFICATION_REF[inputModel.get('quantification')]) {
+        inputModel.set('quantification', QUANTIFICATION_REF[inputModel.get('quantification')], { silent: true });
+      }
+
+      var inputTypeView = new InputTypeView({
+        model: inputModel,
+        columns: this.options.columns,
+        query: this.options.query,
+        configModel: this.options.configModel,
+        userModel: this.options.userModel,
+        modals: this.options.modals,
+        imageEnabled: this.options.imageEnabled,
+        editorAttrs: this.options.editorAttrs ? this.options.editorAttrs[type] : {},
+        disabled: this.options.editorAttrs && this.options.editorAttrs.disabled
+      });
+
+      inputTypeView.bind('click', this._onInputClick, this);
+      this.$('.js-content').append(inputTypeView.render().$el);
+    }, this);
+
+    this._inputCollection.bind('onInputChanged', this._onInputChanged, this);
+  },
+
   removeDialog: function () {
     this._inputCollection.unselect();
     this._fillDialogView.clean();
     this._popupManager.untrack();
   },
 
-  focus: function () {
-    if (this.hasFocus) return;
-    this.$('.js-fillInput').focus();
+  _onInputChanged: function (model) {
+    this.trigger('onInputChanged', model);
   },
 
-  blur: function () {
-    if (!this.hasFocus) return;
-    this.$('.js-fillInput').blur();
+  _onInputClick: function (inputModel) {
+    if (inputModel.get('selected')) {
+      this.removeDialog();
+      return;
+    }
+
+    inputModel.set('selected', true);
+    this._fillDialogView.model.set('createContentView', inputModel.get('createContentView'));
+    this._fillDialogView.render();
+    this._fillDialogView.show();
+
+    this._popupManager.append(this._dialogMode);
+    this._popupManager.track();
   },
 
   _removeForm: function () {
