+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/fill-color-input-collection.js
