+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/fill-color.js
@@ -2,8 +2,8 @@ var _ = require('underscore');
 var Backbone = require('backbone');
 var EditorHelpers = require('builder/components/form-components/editors/editor-helpers-extend');
 
-var FillColorSolidView = require('builder/components/form-components/editors/fill/fill-color/fill-color-solid-view');
-var FillColorByValueView = require('builder/components/form-components/editors/fill/fill-color/fill-color-by-value-view');
+var FillColorSolidView = require('builder/components/form-components/editors/fill-color/fill-color-solid-view');
+var FillColorByValueView = require('builder/components/form-components/editors/fill-color/fill-color-by-value-view');
 
 var TabPaneTemplate = require('builder/components/form-components/editors/size/size.tpl');
 var createRadioLabelsTabPane = require('builder/components/tab-pane/create-radio-labels-tab-pane');
@@ -100,6 +100,8 @@ Backbone.Form.editors.FillColor = Backbone.Form.editors.Base.extend({
 
     if (this.model.get('fillColor').range && this._tabPaneTabs.length > 1) {
       this._tabPaneTabs[1].selected = true;
+    } else {
+      this._tabPaneTabs[0].selected = true;
     }
 
     this._tabPaneView = createRadioLabelsTabPane(this._tabPaneTabs, tabPaneOptions);
@@ -169,6 +171,29 @@ Backbone.Form.editors.FillColor = Backbone.Form.editors.Base.extend({
     }, this);
 
     return this._fillColorByValueView;
+  },
+
+  _onChangeTabPaneViewTab: function () {
+    var selectedTabPaneName = this._tabPaneView.getSelectedTabPaneName();
+    var attrsToUnsetIfFixed = ['domain', 'bins', 'attribute', 'quantification'];
+    var attrs = this.model.get('fillColor');
+
+    if (selectedTabPaneName === SOLID) {
+      _.each(attrsToUnsetIfFixed, function (attr) {
+        attrs[attr] = null;
+      }, this);
+
+      if (attrs.fixed) {
+        if (attrs.range) {
+          attrs.fixed = attrs.range[0];
+          attrs.range = null;
+        }
+      }
+
+      var fillColor = _.extend(this.model.get('fillColor'), attrs);
+      this.model.set('fillColor', fillColor);
+      this.trigger('change', this.model, this);
+    }
   },
 
   getValue: function (param) {
