+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/inputs/input-color-by-value.js
@@ -5,7 +5,7 @@ var CoreView = require('backbone/core-view');
 
 var InputColorValueContentView = require('builder/components/form-components/editors/fill/input-color/input-color-value-content-view');
 var TipsyTooltipView = require('builder/components/tipsy-tooltip-view');
-var InputColorTemplate = require('builder/components/form-components/editors/fill/fill-color/inputs/input-color-by-value.tpl');
+var InputColorTemplate = require('builder/components/form-components/editors/fill-color/inputs/input-color-by-value.tpl');
 
 var Utils = require('builder/helpers/utils');
