+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/inputs/input-color-solid.js
@@ -5,7 +5,7 @@ var CoreView = require('backbone/core-view');
 
 var ColorPickerView = require('builder/components/form-components/editors/fill/color-picker/color-picker');
 var TipsyTooltipView = require('builder/components/tipsy-tooltip-view');
-var InputColorTemplate = require('builder/components/form-components/editors/fill/fill-color/inputs/input-color-solid.tpl');
+var InputColorTemplate = require('builder/components/form-components/editors/fill-color/inputs/input-color-solid.tpl');
 
 var Utils = require('builder/helpers/utils');
