+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/inputs/input-image.js
