+++ lib/assets/javascripts/builder/components/form-components/index.js
@@ -40,7 +40,7 @@ require('./editors/sortable-list.js');
 require('./editors/legend/category-item.js');
 require('./editors/slider/slider.js');
 require('./editors/fill/fill.js');
-require('./editors/fill/fill-color/fill-color.js');
+require('./editors/fill-color/fill-color.js');
 require('./editors/size/size.js');
 require('./editors/taglist/taglist.js');
 require('./editors/datetime/datetime.js');
